style: bottomNav 수정완료
- Nav바 패딩 수정
