+++ src/components/MyPageIcon.tsx
@@ -1,23 +1,25 @@
 import MyPageImage from "@/assets/MyPageImage.png";
 import MyPageActiveImage from "@/assets/MyPageActiveImage.png";
+import { Link } from "react-router-dom";
 
 interface MyPageProps {
   isActive: boolean;
-  onChange: () => void;
 }
 
 const MyPageIcon = (props: MyPageProps) => {
-  const { isActive, onChange } = props;
+  const { isActive } = props;
 
   return (
-    <div onClick={onChange} className="flex items-center justify-center">
-      <img
-        src={isActive ? MyPageActiveImage : MyPageImage}
-        className="h-11 w-auto"
-        loading="lazy"
-        alt="MyPageIcon"
-      />
-    </div>
+    <Link to="mypage">
+      <div className="flex items-center justify-center">
+        <img
+          src={isActive ? MyPageActiveImage : MyPageImage}
+          className="h-11 w-auto"
+          loading="lazy"
+          alt="MyPageIcon"
+        />
+      </div>
+    </Link>
   );
 };
