+++ src/pages/home/index.tsx
@@ -1,3 +1,4 @@
+import BottomNav from "@/components/BottomNav";
 import Nav from "@/components/Nav";
 
 const Home = () => {
@@ -9,6 +10,9 @@ const Home = () => {
       <main className="flex h-full flex-col items-center justify-start gap-10 px-4">
         hello
       </main>
+      <header className="sticky bottom-0 left-0 z-10 border-t-[1px] border-solid border-brandGray bg-white py-2">
+        <BottomNav />
+      </header>
     </main>
   );
 };
